fix: 완료 전단계에서 수파베이스 정보update
